refactor(logger): move interfaces to types.ts inside plugin
